Implement support for scripting terminal apps.
A terminalModuleHandler is implemented.
it initializes terminal modules at startup, allows for terminal modules to receive events and scripts from NVDA but only in terminals.
it allows the user to select a terminal module with NVDA+space in a terminal.
A "test" terminal module was included, a "vim" module for testing will soon be included.